+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.3.0] - 2019-05-15
+### Changed
+- AdwordsId to ConversionId in Settings
+- Sends more product information in productView event
+### Added
+- Includes optional ConversionLabel in Settings
+- Adds orderPlaced event handling
 
 ## [1.2.0] - 2019-05-06
 ### Added